Refactor root analysis class, ready to then build in more sharing/common look-and-feel of controls in particular
