+++ src/analyze/graph/graph_analyzer.py
@@ -3,25 +3,16 @@ from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
 from matplotlib.figure import Figure
 from src.log.log_meta import LogMeta
 from matplotlib import style as mpl_style
+from src.analyze.core.analyzer import Analyzer
 
 
 
-class GraphAnalyzer:
+class GraphAnalyzer(Analyzer):
 
     def __init__(self, guru_parent_redraw, matplotlib_canvas :FigureCanvasTkAgg, control_frame :tk.Frame):
-        self.guru_parent_redraw = guru_parent_redraw
+        super().__init__(guru_parent_redraw, control_frame)
         self.matplotlib_canvas = matplotlib_canvas
         self.graph_figure :Figure = matplotlib_canvas.figure
-        self.control_frame = control_frame
-
-        self.current_track = None
-        self.filtered_episodes = None
-        self.all_episodes = None
-        self.action_space = None
-        self.action_space_filter = None
-        self.log_meta = None
-        self.evaluation_phases = None
-
         mpl_style.use("seaborn")
         self.graph_figure.patch.set_facecolor('lightgrey')
 
@@ -30,43 +21,6 @@ class GraphAnalyzer:
 
     def uses_track_graphics(self):
         return False
-
-    def take_control(self):
-
-        for widget in self.control_frame.winfo_children():
-            widget.destroy()
-
-        self.build_control_frame(self.control_frame)
-
-        tk.Label(self.control_frame, text="                                                 ").pack()
-
-        self.control_frame.pack(side=tk.RIGHT)
-
-    def set_track(self, current_track):
-        self.current_track = current_track
-        self.warning_track_changed()
-
-    def set_all_episodes(self, all_episodes):
-        self.all_episodes = all_episodes
-        self.warning_all_episodes_changed()
-
-    def set_log_meta(self, log_meta :LogMeta):
-        self.log_meta = log_meta
-
-    def set_evaluation_phases(self, evaluation_phases):
-        self.evaluation_phases = evaluation_phases
-
-    def set_filtered_episodes(self, filtered_episodes):
-        self.filtered_episodes = filtered_episodes
-        self.warning_filtered_episodes_changed()
-
-    def set_action_space(self, action_space):
-        self.action_space = action_space
-        self.warning_action_space_changed()
-
-    def set_action_space_filter(self, action_space_filter):
-        self.action_space_filter = action_space_filter
-        self.warning_action_space_filter_changed()
 
     def redraw(self):
         old_axes = self.graph_figure.get_axes()
@@ -77,39 +31,13 @@ class GraphAnalyzer:
 
         self.matplotlib_canvas.draw()
 
+
+    ##########################
+    ### ABSTRACT INTERFACE ###
+    ##########################
+
+    # These are ADDITIONAL to the interface in Analyzer
+
     def add_plots(self):
         # You *MUST* override this
-        pass
-
-    def build_control_frame(self, control_frame):
-        # You *MUST* override this
-        pass
-
-    def recalculate(self):
-        # You MIGHT override this
-        pass
-
-    def warning_track_changed(self):
-        # You MIGHT override this to manage cached or pre-calculated data structures
-        # Do not override to redraw() since Guru already calls redraw() at the right times!
-        pass
-
-    def warning_filtered_episodes_changed(self):
-        # You MIGHT override this to manage cached or pre-calculated data structures
-        # Do not override to redraw() since Guru already calls redraw() at the right times!
-        pass
-
-    def warning_all_episodes_changed(self):
-        # You MIGHT override this to manage cached or pre-calculated data structures
-        # Do not override to redraw() since Guru already calls redraw() at the right times!
-        pass
-
-    def warning_action_space_changed(self):
-        # You MIGHT override this to manage cached or pre-calculated data structures
-        # Do not override to redraw() since Guru already calls redraw() at the right times!
-        pass
-
-    def warning_action_space_filter_changed(self):
-        # You MIGHT override this to manage cached or pre-calculated data structures
-        # Do not override to redraw() since Guru already calls redraw() at the right times!
         pass
